requirements.txt
anyio==3.6.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
cssutils==2.6.0
emails==0.6
exceptiongroup==1.1.1
fastapi==0.95.0
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
premailer==3.10.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pytest==7.2.2
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
zipp==3.15.0
